Logica de alta de producto

+++ Domain/Entities/Orden.cs
@@ -8,13 +8,11 @@ using System.Threading.Tasks;
 
 namespace Domain.Entities{
     public class Orden
-    {
-     
+    {     
         public Guid OrdenId { get; set; }
         public Guid CarritoId { get; set; }
         public DateTime Fecha { get; set; }
-        public int Total { get; set; }
-       
+        public double Total { get; set; }  
         public virtual Carrito Carrito { get; set; }
     }
 }
